cilium-cli: extract portforward logic into new PortForwarder type
Add a new type that provides helper methods for kubernetes
port-forwarding which is not linked to the cilium-cli k8s
client to allow re-use.
